+++ CHANGELOG.md
@@ -20,11 +20,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to measure test coverage with the TROPIC01 model.
 - Documentation: section **Default Pairing Keys for a Secure Channel Handshake** in Get Started
 - GitHub action to run examples against TROPIC01 model (only the supported ones).
+- Replaced `TR01_L3_RES_SIZE_SIZE` and `TR01_L3_CMD_SIZE_SIZE` with `TR01_L3_SIZE_SIZE`.
+- Renamed `TR01_L3_CYPHERTEXT_MAX_SIZE` to `TR01_L3_CIPHERTEXT_MAX_SIZE`.
+- Renamed `LT_L2_DATA_LEN_ERROR` to `LT_L2_RSP_LEN_ERROR`.
+- New generic size macros: `TR01_L3_RESULT_SIZE`, `TR01_L3_CMD_CIPHERTEXT_MAX_SIZE`, `TR01_L3_RES_CIPHERTEXT_MAX_SIZE`.
+- Size macros for L3 results.
+- New return values (`lt_ret_t`): `LT_L3_RES_SIZE_ERROR`, `LT_L3_BUFFER_TOO_SMALL`.
+- L3 buffer size check  to `lt_init` and internal functions.
 
 ### Fixed
 - `lt_ex_show_chip_id_and_fwver`: reboot back to Application mode in the end.
 
 ### Removed
+- `TR01_L3_ID_SIZE` (redundant to `TR01_L3_CMD_ID_SIZE`).
 
 ## [2.0.1]